Support JPEG/PNG to embed ICC Profile if exists into PDF Abstract the ICC parsing into a separate ICCProfile class
