+++ lib/image/png.js
@@ -1,9 +1,10 @@
 import zlib from 'zlib';
 import PNG from 'png-js';
-
+import ICCProfile from './icc_profile'
 class PNGImage {
   constructor(data, label) {
     this.label = label;
+    this.data = data;
     this.image = new PNG(data);
     this.width = this.image.width;
     this.height = this.image.height;
@@ -11,7 +12,7 @@ class PNGImage {
     this.obj = null;
   }
 
-  embed(document) {
+  embed(document, {embedICCProfile=true}={}) {
     let dataDecoded = false;
 
     this.document = document;
@@ -44,7 +45,15 @@ class PNGImage {
     }
 
     if (this.image.palette.length === 0) {
-      this.obj.data['ColorSpace'] = this.image.colorSpace;
+      let colorSpace = this.image.colorSpace;
+      if(embedICCProfile){
+        let iccProfile = ICCProfile.extractFromPNG(this.data);
+        if(iccProfile){
+          iccProfile = new ICCProfile(iccProfile);
+          colorSpace = iccProfile.embed(document, colorSpace, this.image.colors);
+        }
+      }
+      this.obj.data['ColorSpace']=colorSpace;
     } else {
       // embed the color palette in the PDF as an object stream
       const palette = this.document.ref();
